requirements.txt
PySastrawi
transformers
nltk
wordcloud
pandas
numpy
plotly
tqdm